% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_folder.R
\name{clean_folder}
\alias{clean_folder}
\title{Delete files from a folder}
\usage{
clean_folder(path)
}
\arguments{
\item{path}{location of the folder in which to delete all contents.}
}
\value{
NULL.
}
\description{
Function to clean out all the contents of a folder once finished with it.
}
