% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{nvpoint}
\alias{nvpoint}
\alias{nevada_cities}
\title{Nevada cities and roads.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 8 rows and 2 columns.
}
\usage{
nevada_cities
}
\description{
These two sf objects provide outlines for major cities and roads in Nevada
and are used for plot annotations.
}
\keyword{datasets}
