% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{nvshape}
\alias{nvshape}
\alias{nevada_sf}
\alias{nevada_sf2}
\alias{nevada_buffer}
\alias{nevada_buffer_big}
\alias{nevada_roads}
\title{Nevada State Shapefiles}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 17 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 7 rows and 2 columns.
}
\usage{
nevada_sf

nevada_sf2

nevada_buffer

nevada_buffer_big

nevada_roads
}
\description{
These four sf objects are different variations shapefiles for Nevada.
The "buffer" and "buffer_big" are buffers that allows us to fully resolve
the edges of the state when cropping and reprojecting raster values.
The "sf" and "sf_2" values provide outlines for the state and county
endlines for Nevada.
}
\keyword{datasets}
