% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwp_calc.R
\name{pwp_calc}
\alias{pwp_calc}
\title{PWP threshold calculation}
\usage{
pwp_calc(
  ras_list,
  precip_thresh = c(5, 10.1, 20.3),
  temp_thresh = c(0.4, 1.4, 2.7, 4.1)
)
}
\arguments{
\item{ras_list}{A list of rasters, where the first raster is precipitation
forecasts and the second is temperature forecasts. All other elements
of the list will be ignored.}

\item{precip_thresh}{A vector of 6 hour precipitation total thresholds
for the decision support tool (must be exactly THREE thresholds)}

\item{temp_thresh}{A vector of 6 hr max temperature thresholds for the
decision support tool (must be exactly FOUR thresholds)}
}
\value{
A single raster layer of PWP, where the values correspond to
different colors in the visual.
}
\description{
A function to calculate the Present Weather Potential (PWP)
thresholds as contained in Heggli (2023).
}
