% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snodas_prep.R
\name{snodas_prep}
\alias{snodas_prep}
\title{Calculate SNODAS snowpack density}
\usage{
snodas_prep(swe_path, snowd_path, swe_thresh = 10)
}
\arguments{
\item{swe_path}{The full file path to the snodas raster containing SWE
values. Must include the file extension (e.g., tif, nc, etc.)}

\item{snowd_path}{The full file path to the snodas raster containing snow
depth values. Must include the file extension (e.g., tif, nc, etc.)}

\item{swe_thresh}{The amount of SWE required for the ratio calculation
to be relevant. All locations with SWE below this value are assigned
a ratio of zero automatically.}
}
\value{
A single layer spatraster containing the snow density percentages
as needed for the ROS TWI decision support tool.
}
\description{
A function to determine the density of the snowpack, as calculated from
SNODAS gridded products. Note that the function also excludes
all snowpack from consideration if the SWE is less than 10mm
per the recommendations from Musselmann et al. (2019) regarding rain on
snow events with flood generation potential.
}
