% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twi_calc.R
\name{twi_calc}
\alias{twi_calc}
\title{ROS TWI threshold calculation}
\usage{
twi_calc(dens_ras, pwp_ras, dens_thresh = c(26.5, 27.1, 33, 38.1))
}
\arguments{
\item{dens_ras}{A spatraster object with exactly one layer, representing the
density of the snow, expressed as a percentage of the snow water equivalent
to the snowpack depth.}

\item{pwp_ras}{A multilayer spatraster object, where each layer represents
the present weather potential (PWP) for a specific 6 hour forecast window}

\item{dens_thresh}{A vector of snowpack densities, expressed as a
percentage related the snow water equivalent to the snowpack depth.}
}
\value{
A multilayer spatraster object, where the values correspond to the
colors in the ROS TWI decision support tool in Heggli (2023).
}
\description{
A function to calculate the Rain on Snow (ROS) Terrestrial Water Input (TWI)
thresholds as contained in Heggli (2023).
}
